+++ src/settings/InventoryRecordTypeSelectField.js
@@ -3,16 +3,27 @@ import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import { FieldSelect, FieldSelectFinal } from '@folio/stripes-acq-components';
-import { INVENTORY_RECORDS_TYPE_FOR_SELECT } from '../components/POLine/const';
+import {
+  INVENTORY_RECORDS_TYPE_FOR_SELECT,
+  OPTION_WITH_BINDARY_ACTIVE,
+} from '../components/POLine/const';
 
-const InventoryRecordTypeSelectField = ({ label, isRedux, ...rest }) => {
+const InventoryRecordTypeSelectField = ({
+  label,
+  disabled,
+  isRedux,
+  isBindaryActive = false,
+  ...rest
+}) => {
   const FieldComponent = isRedux ? FieldSelect : FieldSelectFinal;
+  const dataOptions = isBindaryActive ? OPTION_WITH_BINDARY_ACTIVE : INVENTORY_RECORDS_TYPE_FOR_SELECT;
 
   return (
     <FieldComponent
-      dataOptions={INVENTORY_RECORDS_TYPE_FOR_SELECT}
+      dataOptions={dataOptions}
       fullWidth
       label={<FormattedMessage id={label} />}
+      disabled={disabled || isBindaryActive}
       {...rest}
     />
   );
@@ -22,6 +33,7 @@ InventoryRecordTypeSelectField.propTypes = {
   label: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
   disabled: PropTypes.bool,
+  isBindaryActive: PropTypes.bool,
   isRedux: PropTypes.bool,
 };